refactor: proxy service readability improvement
optimize getRates method from api call to db call
