+++ src/services/CurrencyService.ts
@@ -1,8 +1,9 @@
 import { getCurrenciesTemplate } from '@windingtree/win-commons/dist/currencies';
 import { CurrencyMeta } from '@windingtree/win-commons/dist/types';
-import proxyService from './ProxyService';
 import currencyRepository from '../repositories/CurrencyRepository';
 import { CurrencyResponse } from '@windingtree/glider-types/dist/win';
+import axios from 'axios';
+import { simardJwt, simardUrl } from '../config';
 
 export class CurrencyService {
   public async getCurrencies(): Promise<CurrencyResponse> {
@@ -29,7 +30,7 @@ export class CurrencyService {
   public async upsertCurrenciesRates(): Promise<void> {
     const currencies: CurrencyMeta = getCurrenciesTemplate();
 
-    const rates = await proxyService.getRates(Object.keys(currencies));
+    const rates = await this.getRates(Object.keys(currencies));
     for (const currency in currencies) {
       const rate: number = rates[currency];
 
@@ -41,6 +42,29 @@ export class CurrencyService {
         decimals: currencies[currency].decimals
       });
     }
+  }
+
+  public async getRates(
+    currencies: string[]
+  ): Promise<{ [k: string]: number }> {
+    const rates = {};
+    for (const currency of currencies) {
+      const ratesData = {
+        source: currency,
+        target: 'USD'
+      };
+      try {
+        const quoteRes = await axios.get(`${simardUrl}/rates`, {
+          params: ratesData,
+          headers: { Authorization: `Bearer ${simardJwt}` }
+        });
+        rates[currency] = quoteRes.data.rate;
+      } catch (e) {
+        // error
+      }
+    }
+
+    return rates;
   }
 }
