+++ src/services/handlers/helpers.ts
@@ -1,0 +1,8 @@
+import { SearchBody } from '../../types';
+
+export type HandlerServiceConfig = {
+  provider: string;
+  searchBody: SearchBody;
+  requestHash: string;
+  sessionId: string;
+};
